# api/swagger/v1/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hospitalApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Staff": {
      "get": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_GetAllStaffs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/Staff/{id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.StaffOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.StaffInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Staff"
        ],
        "operationId": "Staff_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Storage": {
      "get": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_GetAllMedicationStorages",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/Storage/{id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/hospitalApi.DTOs.Outputs.MedicationStorageOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/hospitalApi.DTOs.Inputs.MedicationStorageInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Storage"
        ],
        "operationId": "Storage_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "hospitalApi.DTOs.Inputs.MedicationStorageInputDto": {
        "type": "object",
        "properties": {
          "fkMedicationId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "hospitalApi.DTOs.Inputs.StaffInputDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "lastname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "fkRoleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "hospitalApi.DTOs.Outputs.MedicationStorageOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fkMedicationId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "hospitalApi.DTOs.Outputs.StaffOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "lastname": {
            "maxLength": 100,
            "minLength": 2,
            "type": [
              "null",
              "string"
            ]
          },
          "fkRoleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}